remove malfunctioning unit tests (closes #64, lol)
This commit also adds a global (window) variable called diva, for
testing the public functions through the console (they check out).
It seems that the unit tests just aren't working. The functions themselves
are fine. Chalking it up to a bug in the way I'm writing the unit tests.
How do you test a unit test? Is there a testing framework for QUnit?
Man, unit tests all the way down