Reflow pages upon fullscreen grid resize
Closes #77. Some of the code in loadGrid() is moved out into the helper
function loadGridImages() so that it can be called after resizing. A new
function called resetPanelDims() has been created to take care of updating
settings.panelWidth/Height, taking into account the width of the scrollbar
(which fixes a minor bug that was present previously in fullscreen grid
view).

This reflowing may prove to be slow or annoying (in that it results in more
images downloaded than necessary). In that case, the function can be modified
to use a timer of ~100 ms or so (or more, or less) when deciding whether or
not to reflow. A second timer would be needed if this were to happen, and the
function would have to be called only when in grid view.