Add option to download images (#85)
This works even with JPEG2000 images, despite IIPImage's propensity to
return images at specific dimensions only (when using WID/HEI), as those
specific dimensions happen to be identical to the dimensions we want.

Currently opens a new window, as forcing the image to be downloaded by
the browser doesn't seem easily achievable using Javascript. If necessary,
a PHP script that serves the image with `content-disposition: attachment`
in the header could be added.

The image is temporary and will probably be changed in a later commit.