Move inline-styled background images back to CSS
Some icons were styled in diva.js because they depended on the iconPath
setting. Since there is now an @iconPath variable in variables.less,
this is no longer necessary.

This commit also changes .center-icon to .plugin-icon and fixes that
(temp fix - there should be a better way to do it). A plugin directory
under img/ was also created.

To do:

* Move all the inline styling out of diva.js and do it with
  classes instead
* CSS sprites (requires .plugin-icon rewrite)
* Fix inconsistent iconPath expectations - with or without trailing /?