Modify AJAX request behaviour of diva.js
The document viewer now makes only one AJAX request to the divaserve
script. All the relevant data are sent and/or calculated during this
initial load, and some new functions and settings have been created
to deal with this change.

This commit includes several other related (or not related) changes:

* Some whitespace error fixes
* Check if a tile is horizontally within the viewport (in addition to
  being vertically within the viewport) before displaying it
* More public functions
* The timer in loadPage has been modified to load the page container
  before attempting to load the tiles, resulting in the shadow and
  loading indicator always appearing (which is the expected behaviour)
* Added some functions to validate zoomLevel and pagesPerRow

Tests and changes in other files forthcoming.