Add white outline to fullscreen icon
Otherwise it gets lost against a dark background (e.g., when we're in
fullscreen mode and there's a really dark page). It still doesn't look
great in this case but it definitely doesn't look worse (and the white
outline looks kind of nice against a light grey background, too).

The size of this icon is now configurable via the @fullscreenIconSize
variable in case it ever has to be changed again.