Partially fix minor styling bug in IE9
In IE9, it used to be that when you scroll through the document and
consequently change pages, the page number label in the top right
corner would sometimes span two lines. Namely, if the current page
number is 2 digits or more, it will sometimes (but not always) push
the number of pages element down into the next line. This is very
strange and doesn't seem to be triggered in a deterministic fashion,
as sometimes I can scroll through the entire document without having
it happen whereas other times it will happen as soon as I hit page 10.
Also, whenever I uncheck then recheck any style attribute from the
"Trace Styles" pane in the developer tools window, the problem
magically goes away. I tried to simulate this programmatically but
nothing seems to work. I don't even know

The fix introduced in this commit forces the page number label to span
only one line. This has the side-effect of making it sometimes encroach
upon the grid icon's territory, but, what can you do.

Notably, the repainting is done properly in IE8, which is pretty odd.
This isn't an issue in any other browser that I've tested (Opera, FF,
Chrome) so it seems that this is just IE being IE

For issue #94.