Fix IE8 compatibility issue with loading pages
window.innerHeight/Width are not supported in IE8 and older (but we
don't need to support older). However, they should be equivalent to
document.documentElement.clientHeight/Width, so if the inner properties
aren't available (by way of being undefined) then we have a fallback,
and everything loads nicely in IE8. Although there are no drop shadows,
which is a shame, because drop shadows make everything look nicer.

For issue #94. Almost closed.