Add ability to disable plugins in own init fn
Now, plugins can be disabled programmatically, by the plugin itself.
This is necessary for the canvas plugin to automatically disable itself
if it detects that the user is using a browser that does not support
canvas (notably, IE8).

This is the last of the IE-compatibility-checking commits. So, without
further ado, I can close #94.