Show loading indicator during slow ajax requests
The styling for the loading indicator originally created for the canvas
plugin was backported into the main diva stylesheet. Now, the throbber
is created by the diva instance and can be used by plugins. This commit
also makes the canvas throbber timeout configurable via the settings.
Same for the initial diva throbber timeout.

Closes #109.