New mandatory setting parameter: imageRoot
This parameter corresponds to the image root directory set in the divaserve file.

Note: By default, this is relative to your root directory, so it's set to /something/somethingElse. However, you can set the FILESYSTEM_PREFIX for IIP to provide a new root. If you have set this, imageRoot should be relative to the directory you defined in FILESYSTEM_PREFIX. If your images are in /mnt/images/projectA/foo.jp2, and you have set FILESYSTEM_PREFIX to /mnt/images, you should set imageRoot to projectA; If you have not set FILESYSTEM_PREFIX, you should set it to /mnt/images/projectA.