New: enableGridControls and enableZoomControls settings
Replaced enableGridSlider and enableZoomSlider settings with enableGridControls and enableZoomControls. This is less complicated than the alternative of having 4 settings (enableZoomSlider, enableZoomButtons, enableGridSlider, enableGridButtons).
This does assume that there are no usage cases in which both buttons and slider should be displayed at the same time (cumbersome with current layout).

See discussion on #130