Fixed: margin in fullscreen cutting off scrollbars
In fullscreen, there was a margin-left of settings.viewerWidthPadding being applied every time browser dimensions were adjusted (within adjustBrowserDims()). This is not desired behavior in fullscreen (and adding the style every time the window size changes serves no purpose). To solve this, within handleModeChange(), we instead set the margin defined by viewerWidthPadding if not in fullscreen, and one of 0px if entering fullscreen.

Fixes #137