Fixed: Window size smaller than initial CSS height caused outer page to scroll
When the browser size is smaller than the window size, set the heightProportion to a default of 0.9 instead of using a value greater than 1, which causes overflow.

Note: this fix only works when enableAutoHeight is set to true. This is proper behavior because if enableAutoHeight is false, the user will set a static height that should not be changed regardless of whether it is larger than the window..

Fixes #141