Fixed: viewerWidthPadding margins still present when enableAutoWidth is off
According to the code documentation for viewerWidthPadding, the margin should not be present unless enableAutoWidth is on. This makes sense, as when someone is embedding Diva into a webpage, they would probably prefer to control Diva via CSS, just as they are controlling the presentation of all the other elements on the page.

Refs #147