New: CORS error handler for remote JSON
When an unknown error occurs, we check whether the path to the JSON file (objectData) is absolute (if it is a relative path, a CORS error is impossible.) If it is an absolute path, we extract the hostname and compare it with the hostname of the current page. If they do not match, it is very likely that the error is CORS-related. In this instance, we inform the user of the nature of the error via the error box and link them to the relevant documentation.

Closes #149