New: verticalOffset calculated from center of page
Within the getState() function (used for getting a shareable link) and gotoPage function (called upon document load), verticalOffset is now calculated from the center of the relevant page. This sufficiently allows a view of a page to be easily communicated across browsers without having to resize the receiving user's document viewer.

Fixes #151