New: translateFromMaxZoomLevel method
This method will translate a set of co-ordinates in reference to the maximum zoom level and will translate them to co-ordinates on the current zoom level.

For example, if a reference is 1000 on the image (could either be a pixel co-ordinate or a width/height) taken from a maximum of 4 zoom levels, this method will translate it to 250 at zoom level 2.

This commit includes a unit test for this method as well.

Fixes #156