Fixed: uneven margins when enableAutoWidth enabled
On the iPad, the right margin was smaller with enableAutoWidth enabled. This is because window.innerWidth on the iPad does not include the body margin, and so the body margin was pushing the viewer to the right. To fix this, we negatively x-offset the viewer by the body margin.

Fixes #163