Fixed: throbber now shows for lower zoom levels
showThrobber() guesses from zoom level whether the image will take long to load, and previously only displayed the loading indicator if the zoom level was larger than 2. From my testing (simulating 3G and DSL connections), it would be useful to have the indicator on anything higher than the first zoom level.

Also updated the loading indicator to hide on error and when the user closes canvas view, if it is still present.

Refs #187