New: videos on front features page
Feature videos to replace the previous feature images. Simple Javascript to only play videos when they are fully in the viewport. I tweaked ffmpeg a bit to compress the videos as much as possible without significantly compromising quality. Will test with simulated slower connections to see if videos need to be optimised further.