Fixed: redundant button CSS
Previously, the LESS stylesheet was generating CSS for buttons using mixins, meaning that the generated stylesheet included many repeated styles. Now, we just create a single 'button' class inside mixins.less and apply it using the 'class' attribute of the HTML tags. This means the 'button' chunk of CSS now only gets loaded once, instead of for each button that uses it.