Fixed: plugin Icon classes hardcoded
I was going to replace the class name filter with one based on the parent, but it seems that it is no longer needed to filter elements to enable events to pass through the jQuery.kinetic element. Tested on iPad, and tapping canvas and download icons works without the filterTarget code.

Fixes #200