Fixed: icon hover state persistent on mobile devices
Previously, tapping on a button on the iPad and subsequently interacting with other page elements caused the buttons to flash due to broken :hover behavior in mobile Safari. This CSS media query makes sure that hover effects are not applied to devices with a screen width of <=768 pixels (i.e. the iPad and other tablets). (Retina iPads still report a device width of 768 pixels, in portrait or landscape.)

Fixes #159