Fixed: Rounding bug garbles images at specific page widths
If an original image dimension was an exact multiple of 256, then each of the smaller dimensions would be calculated as sub-1 pixel width. Previously, the generate_json.py calculations would round that up to the next nearest column, but this would result in a wrong calculation for the numbers of rows and columns present in the image.

This commit fixes this by looking at the remainder of a dimension. If it is less than one pixel it will round down to the closest whole number dimension; otherwise, it will keep with the previous behaviour and round up.

Fixes #243