Fixed: extra plugin icons breaking fullscreen tools UI
Previously, the addition of the autoscroll icon to the toolbar meant that in fullscreen grid mode, the diva-right-tools spilled off of the toolbox because the 'Pages per row' label is wider than the 'Zoom level' label.

Now, we don't declare a width so that the toolbar size changes dynamically based on the width of the items within it, so the layout is never broken. (CSS overhauls coming for version 4.0).