New: settings.isIIIF
settings.isIIIF replaces settings.iiifServerURL. settings.iiifServerURL is superfluous because if the user wants to use IIIF they can simply supply a IIIF-compliant Manifesto JSON resource to settings.objectData. (This is assuming that the two possible use cases are using IIPImage syntax and the generate_json JSON file; or using an IIIF Image server and IIIF Manifesto.) The thought is that users can just as easily generate a IIIF manifesto as they can a generate_json-style one if we include a script with Diva releases (to do).