WIP: changeObject(objectData) public function
I refactored the existing $.ajax.success function to break out parsing of data into parseObjectData, so that I can also parse the data on the request made in changeObject.

I need to run the tests and make sure that this doesn't break anything, and make sure that I'm only making the necessary calls in changeObject.