WIP: Openings/book layout
Starting implementing openings in document view. For now, the first page will be displayed as a cover on the right, and the following pages will be displayed facing, unless they have a viewingHint of 'non-paged', in which case they are excluded when using book layout.

* Added 2 settings: inBookLayout and documentPaged.

inBookLayout is set to true when the user is viewing the document in openings mode.

documentPaged is automatically set when the sequence viewingHint is 'paged'. If inBookLayout is true while documentPaged is false, an error will occur.

* settings.pages[i].paged is false when the corresponding canvas in the manifest has a viewingHint of 'non-paged', meaning that it should be excluded from Book layout.