Fixed: document chopped off at bottom
The total height of the document is now calculated properly in book view (on the fly instead of in parseManifest, since we have the data we need as a side effect of calculating the page layout). The inner element is now the right size to fit all of the page images.