Fixed: pages not appearing properly in book view
In book layout, pages were not appearing if the user was zoomed in to the point where one page is visible and then scrolled up or down.

This is because attemptPageShow was attempting to show only the page immediately preceding or following the current page. However, when we're zoomed into the page with index 2 and scrolling down, we need to consider whether or not to show page 4, because it is the next page that will be encountered when scrolling (because both pages 2 and 4 are on the right-hand side.)

Fixed by evaluating whether the next two pages should be shown when scrolling.