New: cropping via regions supplied in IIIF manifest
The user can now specify regions of images to be displayed on a canvas in a IIIF manifest. In order to do so, the user includes coordinates in the format x, y, width, height in the value for canvas.resource.@id. For examples, see http://iiif.io/api/presentation/2.0/#segments .

Closes #96