New: Don't add meta elements to the document
Previously, Diva appended some meta elements to the document head
if it detected that it was running on a mobile device. These could
interfere with meta elements already on the page, and setting them
is out of scope for a library that can integrate with other other
content in the same document.