New: placeholder element to fill out first opening in Book Layout
We want openings to be centred by default. However, for the first opening, it looks odd to have a single page on the right half of the viewport and empty space on the left. So we draw a placeholder element that fills out the left half of the screen, showing the user that when they scroll down, they can expect that a page will be there.