New: canvases with a viewingHint of 'non-paged' skipped in book layout view
As specified by the IIIF Presentation API, canvases with the property 'viewingHint: non-paged' are skipped over when displaying book layout. In the parseManifest function, canvases with 'non-paged' gain the property 'paged: false' when added to settings.pages[]. During rendering of document view, if in book layout, we look for this property and exclude these pages from the page layout calculations in calculatePageOffsets(). Also, when the viewer attempts to load a non-paged canvas in loadPage, the function returns early.

Refs #136