Fixed: toggle orientation in grid view results in undefined behavior (/explosions)
:boom:

Previously, calling toggleOrientation in when in grid view broke everything. Now, when toggling orientation in grid view, we switch out of grid view and into either document or book mode (whichever the user was in last) in the new orientation.

Modified changeView() to not disable book view when switching into grid, so that the previous view the user was in is preserved when switching in and out of grid (this change is global).

Fixes #285