Fixed: events scoping works!
Changed the structure of the Events cache: its keys are topics, and their values are objects. Within each topic object, the keys are either 'global' or the ID of the instance the subscribers listen to, and values are an array of subscribed functions to execute.

`diva.Events.subscribe` takes a string representing the ID of the Diva instance to subscribe the function to (optional: if none specified, the function is subscribed to all instances.)

See #234 for details about unsubscribe functions. Will update documentation.