New: tiles rendered on canvases
Instead of being div elements in the DOM, tiles are now drawn into a canvas on each page.

isTileLoaded has not been updated to check whether tiles are loading, so we don't see anything yet (I patched it to return false on my working copy for testing purposes).

TODO: Refactor createDrawTileFunction? It returns a callback that draws each image once it has loaded. It is currently defined inside loadTiles because we need access to the context variable in each callback (the canvas drawing context).