WIP: settings.loadedTiles to track loaded tiles
settings.loadedTiles is an array that keeps track of which tiles have already been loaded, to be used by isTileLoaded. (It replaces the previous method of DOM lookup used by isTileLoaded.) This array is cleared in resetTilesLoaded(), which is subscribed to DocumentWillLoad.

The array is two-dimensional (first dimension is page index, second dimension is tile index.)

Using an array to keep track of loaded tiles is faster than relying on getting pixels from the canvas and checking if a tile is there or not, since context.getImageData() is slow and would need to be called every time we check if a tile is loaded.

WIP because I have experienced very occasional instances where pages fail to load using this method. Working on isolating those.