Fixed: blurred page images
The canvas drawing context must be an integer of pixels wide. When a non-integer value is supplied, canvas items are drawn to a context rounded to the nearest integer size and then stretched to a non-integer size in CSS. This results in significant blurring of the image (there appears to be no good sub-pixel antialiasing when a canvas context is stretched, at least in Chrome.)

Rounding down (Math.floor) yields sharp images.