Refactor: createToolbar
Previously, createToolbar concatenated lots of strings and used jQuery to append them to the DOM.

It is faster (rendering) and cleaner to use the DOM native methods to create and add the toolbar elements. (Even though this results in more lines of code in the Diva core, it is most probably less code than was being executed by jQuery in the previous implementation.)

Functions have also been created for the repeatable elements of the process.

Related to the CSS optimizations, buttons are now span elements (instead of divs) with display: inline-block, since this allows for more canonical control of layout rather than having to rely on floated elements.