New: Basic CommonJS support
This commit adds a wrapper function which is inserted around Diva
at build time. It checks for CommonJS environments and tries to
load a local version of jQuery.

As of this commit Diva will still always expose some globals,
specifically, window.diva and window.divaPlugins. A few things
which *had* been global, but don't seem to need to be are now
local. These are the `activeDiva` and `activeDivaController`
variables in utils.js. (If these actually need to be exposed they
could be attached to the `diva` object.)

Closes #292