WIP: smooth double-click zoom
Changes handleDocumentDoubleClick so that when double-clicking, the clicked point is the point of origin for the zoom, so it maintains position.

The transform transition at the end of handleZoom (the smooth part) requires coordinates relative to the top left of diva-inner, while handleZoom requires the coordinates of the new viewport center relative to the top left of the page, hence the new dblClickX and dblClickY arguments for the transform coordinates.

In handleDoubleClickZoom, compensated for fixed/adaptivePadding to maintain proper center.

loadDocument is now fired when the scale transition event occurs (triggered from handleZoom).

TODO:
- Start preloading tiles as soon as the zoom is initiated by handleZoom
- Smoothly transition in to the new  zoom level in loadDocument, preserving the old tile data instead of flashing blank as it does now