WIP: preloading pages
preloadPages is a new function which currently determines which pages will be in the viewport at a new zoom level. It will soon preload them onto offscreen canvases as the scaling transition (smooth zooming animation) is starting. On transition end, the preloaded (or still loading at that point) offscreen canvases will then replace those that contain the data from the old zoom level.

It was necessary to calculate whether pages that are not yet in the DOM will be visible at the new zoom level, so isPageInViewport spawned a similar function, isPageInViewportBounds. This function behaves the same but takes dimensions via its arguments instead of reading them from the DOM. It might be advisable to continue moving the direction of not reading the DOM for these types of calculations in the future in order to maximize performance and reliability. In that vein, the *Bounds functions will soon replace their progenitors.

loadDocument and gotoPage were modularized, resulting in calculateDocumentDimensions and calculateDesiredScroll, respectively. These functions are used in preloadPages.