Fixed: ctrl-double click zooming out too far on Safari
The dblclick event fires after the contextmenu event when zooming out by ctrl-double-clicking on Safari, but not Chrome. This is not the case on develop, even though no code in this section has changed.

Ignoring the dblclick event when zooming out solves the problem, as the zoom is now triggered once by the contextmenu event.