WIP: temporarily prevent undefined behavior
This is a temporary patch for some unexpected behavior that will be fixed over the coming week.

The changes to createToolbar() throttle the rate at which handleZoom is
called from the zoom buttons, preventing undefined behavior resulting
from handleZoom being called in the middle of a zoom. Ideally (soon
actually), handleZoom will gracefully handle multiple calls within a short
timeframe and dequeue the previous zoom operations.

The changes to clearViewer and handleZoom prevent a similar issue with
zooming with the mouse on the viewer pane - mouse interaction is
disabled until the zoom animation is finished.