Fixed: losing page when zooming out
In Chrome, when zooming out to lower levels (0 or 1), sometimes the viewer would lose the current page. This was because during the zoom transition, scrollFunction was triggered, updating the scroll position (settings.verticalOffset) to an invalid one. Now, we simply unbind the scroll callback so that scroll is not updated between the beginning of zooming (end of handleZoom) and the end (when loadDocument is triggered by the transition end.)