Fixed: unsubscribe not working as expected
JSHint pointed out that the `delete` operator was not working as expected, as the assignment of an array within `cache` to `topicArray` is by value, not by reference; deleting topicArray produced no effects on cache. Instead, we have to write a full reference to the property we want to delete from cache because of this limitation of Javascript.