Fixed: isTileVisibleInBounds merged into isTileVisible
isTileVisible now requires arguments specifying the viewport rather than accessing the DOM directly. This allows us to optimize it by caching the values outside of any loops, and to use it for preloading pages before the viewport is adjusted in the DOM, since it doesn't depend on the DOM. Each argument is specified separately rather than in a single object for performance reasons.