Refactor: Redundant initialization parameter
Previously the jQuery function wrapping the Diva initializer added
a parentObject value to the config object before passing it in to
the Diva instance. This was redundant since the parent element is
*also* passed to the instance, which wasn't doing anything with it.

This commit has the Diva function build settings.parentObject based
on its element argument, without having the jQuery wrapper function
mutate the config object.