New: settings.isZooming
settings.isZooming is a flag to keep track of whether zooming is currently in progress (i.e. if the transition between zoom levels is still running). This information is useful when we are calling handleZoom during zooming and deciding how to update the currently displayed pages (i.e. we will want to preload pages for the most recent zoom only and cancel the previous preloads).