Fixed: setCurrentPage in Book View
Previously, setCurrentPage was not changing the page reliably in book view and vertical orientation. Specifically, scrolling sideways resulted in erratic behavior.

Several changes resulted from assumptions being made that pages with even indexes are on the right, which is incorrect now that we support the IIIF canvas viewingHint property, allowing the user to exclude certain pages from view to preserve the book structure.

The modifications are
(1) in adjustPages, run setCurrentPage even if the scroll is not in the primary direction
(2) new function getClosestVisiblePage, called when considering candidates for the new page to be set
(3) setCurrentPage now checks which side of the document the user is viewing and attempts to switch to the book page corresponding to that side if it is not already set.