New: padding scales with zoom
In order to properly animate zooming without having pages snap back to their new positions upon scale finish, it is necessary that the inter-page margin scales with the zoom level, so this has been implemented in handleZoom.

This will break all of the existing expected measurements of page offsets in the unit tests.

Refs #304