New: settings.previousZoomLevelCanvases
Before, preloadPages was grabbing the page images from page divs in the DOM, but this meant that if the user zoomed again while a zoom was in progress, the page images would be deleted and not redrawn fast enough to prevent blanking if preloading hadn't completed.

previousZoomLevelCanvases is an array that holds page canvases at the initial zoom level, so that preloadPages can grab them and render the data from the old zoom level onto the new canvases before rendering the image data for the new zoom level.

This way, the canvases loaded by preloadPages don't blank in between zooming, even if the user keeps zooming while a zoom is currently in progress.