New: Read hash parameters in one shot
Previously different hash parameters were read and accounted for
at various stages in the initialization, which made the overall
logic difficult to follow and maintain. This commit tries to mitigate
that by consolidating the changes into one place, using the same
logic as setState. This unavoidably (and hopefully harmlessly)
breaks some undocumented assumptions about when values are
available, but the known, documented breaking changes are:

 1. Invalid grid pages per row values set via hash params are
    replaced with the maximum legal value instead of being ignored,
    for consistency with how the zoom level validation works.
    (This can be revisited.)
 2. x/y offset values are ignored unless the page is specified.
    This seems like common sense, and probably won't change anything
    except the hashparam tests.