Fixed: Don't default to global Events.unsubscribe
Previously Events.unsubscribe reverted to unsubscribing from the
global scope if no callbacks were ever registered for the specified
scope. While there are arguably cases where this behaviour might be
helpful, it was undocumented and interacted poorly with the
completely option. Since it's not obvious it would ever be desirable,
I'm just removing it.