New: Rework page offset calculation
Calculate the page offsets based on groups of pages which are laid
out relative to each other, and then position the groups according
to the primary axis. This restricts how much of the offset code
needs to know about the distinction between document and book
view, and will hopefully provide a way to share more code with
grid view in the future.

This commit introduces some minor changes to the book-view offsets
because it rounds values more consistently.