Refactor: Simplify page scroll handling
Simplify the handling of updating which pages are visible.
Previously page visibility checks leveraged a lot of knowledge about
the current view and the direction of scrolling. The revised
implementation knows nothing aboout the positioning and just
iterates through all the page groups to see which are visible.
More optimizations can be added back in if it turns out to be
necessary.