New: Render grid view on single canvas
Replace the specialized div-based grid view implementation with a
canvas implementation using the general layout code. This makes
the single canvas renderer the only rendering implementation.

In the process of doing this I made some changes to the current
page selection logic for grid view, because the old logic was
difficult to reproduce and the desired effect didn't seem to be
tightly specified. I also just dropped some test assertions related
to what pages are visible; I think it should be safe enough to
assume that those things are working at the renderer layer if the
high layer is working okay. (Adding more tests for the renderer in
isolation would be good of course.)

I also moved the code which does image selection to the viewer
core temporarily so that grid view can have a different, non-tiled,
implementation.