Refactor: Separate sized, scroll container element
Previously .diva-outer was the element to size to change Diva's
size, and also the element on which scrollbars appear. A
prerequisite for rendering the document on a single canvas is that
these be separate elements so that the canvas can be placed in a
fixed position outside of the scroll area, but still within Diva's
sized viewer. There are basically two ways to accomplish this:

 1. Create an element which is a parent to .diva-outer which is
    sized, with .diva-outer being sized at 100% its parent's height.

    This breaks code which assumes that .diva-outer is the outermost
    element, or that tries to change Diva's viewport by sizing the
    .diva-outer element.

 2. Create a child of .diva-outer which is sized to 100% of
    .diva-outer's height, and place the scroll region inside it.

    This breaks code which tries to access the viewport position
    via the scroll of .diva-outer.

This commit implements the second option, because while it affects
more code within Diva, it's less likely to break user code. It's
very common to customize the Diva viewer's size, but shouldn't be
necessary to change the positions of its scrollbars directly.