Refactor: Drop dead code from viewport object
When I wrote the viewport code I added a system where you could
mark the viewport state as invalidated but defer the read of the
updated values until a later point. I think I thought I needed
that for zoom animations or something, but it really doesn't seem
like the right approach.