Refactor: Make settings object read-only
Turn the settings object into a read-only facade which exposes the
properties of separate viewerState and options objects.

This is a prerequisite for splitting the object which exposes
a viewer instance's public interface and the underlying viewer
logic: it will allow the externally-exposed settings object to
be mostly maintained for backwards-compatibility while the internal
management of the data is allowed to change.