New: Use lower-res tiles while tiles load
This sizeable commit implements layered tile drawing. To do so, it
turned out to be more convenient to defer the calculation of scaled
tiles until later in the rendering process, so that logic has been
moved from ImageManifest (which once again only gets tiles for
integer zoom levels) to Renderer.