New: Unify release bundle structure
Previously the build linked to on the website and via GitHub
releases and that distributed on npm were different. This commit
moves all the release code to be based on npm's packaging for
consistency. As a result, source files and metadata files are
no longer copied into build/, which now contains only generated
files. This is in keeping with the best-practice structure which
has developed for npm modules, where source and dist files are
placed in separate directories, and will also help with the kind of
confusion seen in #266.

As support for this change, this commit updates the develop:server
gulp command to work with the root directory and improve the checks
on the packaging process.