New: Support for 'non-paged' pages
Pages tagged as 'non-paged' in a IIIF manifest are now
hidden in all views by default. A new button in the toolbar toggles their
visibility.

Here are the changes introduced by this commit:
New public methods:
- toggleNonPagedPagesVisibility()
- showNonPagedPages()
- hideNonPagedPages()

Changes to public methods:
- isValid: Now also checks if the page is paged or if non-paged pages
are visible.

New settings:
- enableNonPagedVisibilityIcon: show or hide the icon in the toolbar that
toggles visibility of non-paged pages. Note that the icon is automatically hidden
if there are no non-paged pages in the current document. Default: true
- showNonPagedPages: Whether pages tagged as ‘non-paged’ (in IIIF manifests only)
should be visible. Default: false

Closes #347, Refs #314