New: Exposed hook for go to page function
New settings:
- onGotoSubmit: When set to a function that takes a string and
returns a page index, this will override the default behaviour
of the 'go to page' form submission. If the function returns
false or null, a message is displayed to the user saying that
the page was not found.
- enableGotoSuggestions: Defaults to true. Controls whether
suggestions are shown when a user is typing in the 'go to page'
form.

Closes #265