New: Diva now fills its parent's height
Structure changes: A new element has been added with
ID 'diva-[id]-wrapper' and class 'diva-wrapper' which
is a container for every element created by Diva.

Settings changes: A new setting 'fillParentHeight' with
a default value of 'true'. The main benefit of turning
this off is that Diva will not use any flexbox layout
which would allow avoiding compatibility issues with
older browsers. When set to false, the .diva-outer default
size is 700px, as it was previously.

Closes #253