Fixed problem with animation not being displayed after a zoom action #382
The problem was that the oldZoomLevel was being set to the current zoom
level then this would be assigned to the initialZoomLevel, which would
return a wrong initial zoom level. This did not cause any problem
before but it might be that ES6 syntax does not behave the same way as
ES5.

Solves Issue #382