New: Add appropriate page label range for view mode
There seems to be an issue with the page coordinates and dimensions not being defined properly, but rather offset vertically (upwards) by around 100-200px. This is not caused by the page indices change, as it is still present before these changes. It is most easily observable by setting the viewportMargin to 0