Fixed: Page tool icons missing on first page
This was caused by the previous fix to the page region being miscalculated - the proper fix for the issue is to add the padding.top to the bottom of the page region such that the height of the plugin icons is accounted for. Also adjusted the unit tests now that the icons are included in the page region.